+++ tests/sdk/nodejs/replace-daemonset/step1/index.ts
@@ -1,0 +1,17 @@
+import * as k8s from "@pulumi/kubernetes";
+
+const appLabels = {
+    app: "nginx",
+};
+
+const daemonset = new k8s.apps.v1.DaemonSet("test-replacement", {
+    spec: {
+        selector: { matchLabels: appLabels },
+        template: {
+            metadata: { labels: appLabels },
+            spec: { containers: [{ name: "nginx", image: "nginx" }] },
+        },
+    },
+});
+
+export const name = daemonset.metadata.name;